+++ lightcurve/src/lightcurve_api/models/lightcurve_model.py
