--- lightcurve/src/magstats_api/README.MD
@@ -1,1 +1,0 @@
-TEST
