+++ lightcurve/src/magstats_api/models/magstats.py
