+++ lightcurve/src/object_api/models/object.py
