+++ lightcurve/src/object_api/routes/rest.py
@@ -2,8 +2,6 @@ from fastapi import APIRouter, HTTPException, Request
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 
-from core.exceptions import ObjectNotFound
-from core.services.object import get_object
 
 router = APIRouter()
 templates = Jinja2Templates(
