+++ lightcurve/src/probability_api/models/taxonomy.py
